refactor(extension): split mod.rs into per-concern files
extension/mod.rs bundled types, sandboxing, the runtime wrapper, the
esbuild bundler, and the SEC101-SEC115 security scanner in one
2713-line file. Split by concern into types/sandbox/runtime_wrapper/
bundler.rs and a security/ module (rule data in rules.rs, detection
tests split by rule range). mod.rs re-exports the same public surface
so no caller needs to change.
